whatsapp_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dateutil==2.8.2
openai==1.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0 

# Google Calendar API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2

# Processamento de linguagem natural
dateparser==1.2.0
spacy==3.7.2

# AI Libraries
openai>=1.0.0
httpx>=0.25.0,<0.28.0  # Versão compatível com OpenAI
google-generativeai>=0.3.0  # Google Gemini API 